completed basic version, tests passing

+++ src/main/kotlin/org/toastedware/hackazon/Cart.kt
@@ -1,7 +1,9 @@
 package org.toastedware.hackazon
 
+import java.math.BigDecimal
+
 data class Cart(
         val products: List<Product>,
-        val salesTaxes: Double,
-        val total: Double
+        val salesTaxes: BigDecimal,
+        val total: BigDecimal
 )
